now inrunODE works, still to fix lento?
and change all the depending app.py and alastulo? , inter?
or whatever there is that still use the old function inrun. Or actually
still to do is to change the name of ratkaise-->inrun and just change
the depending module to inrunODE.

And lento will have to be fixed in the same manner. that should be easy,
fingers crossed...